ZKP-670 changes type from u128 to u64

+++ pallets/prices/rpc/runtime-api/src/lib.rs
@@ -8,9 +8,9 @@ use primitive_types::U256;
 sp_api::decl_runtime_apis! {
 	pub trait PricesApi {
 		fn get_remaining_markets() -> Vec<U256>;
-		fn get_no_of_batches_for_current_epoch() -> u128;
+		fn get_no_of_batches_for_current_epoch() -> u64;
 		fn get_last_abr_timestamp() -> u64;
-		fn get_remaining_pay_abr_calls() -> u128;
+		fn get_remaining_pay_abr_calls() -> u64;
 		fn get_next_abr_timestamp() -> u64;
 		fn get_previous_abr_values(starting_epoch: u64, market_id: U256, n: u64) -> Vec<ABRDetails>;
 	}
